code duplication removed
